backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
selenium==4.15.2
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3
python-dotenv==1.0.0
webdriver-manager==4.0.1
newspaper3k==0.2.8
nltk==3.8.1
spacy==3.7.2
pydantic==2.4.2
httpx==0.25.1
html2text==2020.1.16
pandas==2.1.1
aiohttp==3.11.18
python-multipart==0.0.6
google-generativeai==0.3.1 